+++ flow-typed/gui.js
@@ -2,5 +2,4 @@
 declare type ListInjectedItem = {
   node: Node | (index: number, lastVisibleIndex: ?number, pageSize: ?number) => Node,
   index?: number,
-  replace?: boolean
 };
